requirements.txt
langgraph==1.0.4
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-anthropic==0.3.1
sqlalchemy==2.0.44
numpy==2.3.5
fastapi==0.123.0
uvicorn==0.38.0
python-dotenv==1.2.1
thefuzz==0.22.1
python-levenshtein==0.27.3
websockets==15.0.1
openai==1.58.1

